feat: add structure for rust expression parser
