+++ yarte_rust/src/lib.rs
@@ -1,7 +1,7 @@
-pub mod error;
-pub mod expr;
+mod error;
+mod expr;
 pub mod lexer;
-pub mod literals;
-pub mod parser;
+mod literals;
+mod parser;
 pub mod sink;
 pub mod token_types;
